requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.2.10
astroid==2.4.2
awscli==1.18.69
bcrypt==3.1.7
blinker==1.4
botocore==1.16.19
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1
django-crispy-forms==1.10.0
django-debug-toolbar==3.1.1
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
isort==5.6.4
jmespath==0.9.4
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pipenv==11.9.0
protobuf==3.6.1
pyasn1==0.4.2
pycairo==1.16.2
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyMySQL==0.10.1
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.3
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.0
s3transfer==0.3.3
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.3.1
systemd-python==234
toml==0.10.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
wrapt==1.12.1
xkit==0.0.0
zipp==1.0.0
